+++ dulcecontrol-backend/src/main/java/com/dulcecontrol/bakery/repository/SedeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.dulcecontrol.bakery.entity.Sede;
 
-public interface SedeRepository extends JpaRepository<Sede, Integer> {
+public interface SedeRepository extends JpaRepository<Sede, Long> {
 
 }
